ActuacionView en progreso, se separo en un componente las cards
